+++ views/about.php
@@ -1,13 +1,13 @@
 <h1>About DRW</h1>
 
 <p>DRW (Europe) Ltd, Science Village, Chesterford Research Park, Little Chesterford, CB10 1XL, UK</p>
-<p>Tel: <a class="alt" href="tel:+441223784332">+44 1223 784332</a></p>
-<p>Email: <a class="alt" href="mailto:sales@drw-ltd.com">sales@drw-ltd.com</a></p>
+<p>Tel: +44 1223 784332</p>
+<p>Email: sales@drw-ltd.com</p>
 <br>
 <p>DRW (US) Ltd, 845 Embedded Way, Suite 150, San Jose, CA 95138, USA</p>
-<p>Tel: <a class="alt" href="tel:+14087731511">+1 408-773-1511</a></p>
+<p>Tel: +1 408-773-1511</p>
 <p>Fax: +1 408-773-1553</p>
-<p>Email: <a class="alt" href="mailto:sales@drw-ltd.com">sales@drw-ltd.com</a></p>
+<p>Email: sales@drw-ltd.com</p>
 <p>EIN: 010584612</p>
 <p>DUNS: 141057898</p>
 <p>Diagnostics for the Real World is a company incorporated and registered under the laws of the State of Delaware (USA).</p>